+++ elide-core/src/main/java/com/yahoo/elide/core/RequestScope.java
@@ -245,8 +245,13 @@ public class RequestScope implements com.yahoo.elide.security.RequestScope {
         this.mutatingMultipleEntities = outerRequestScope.mutatingMultipleEntities;
     }
 
+    @Override
     public Set<com.yahoo.elide.security.PersistentResource> getNewResources() {
         return (Set<com.yahoo.elide.security.PersistentResource>) (Set) newPersistentResources;
+    }
+
+    public boolean isNewResource(Object entity) {
+        return newPersistentResources.stream().filter(r -> r.getObject() == entity).findAny().isPresent();
     }
 
     /**