+++ elide-core/src/main/java/com/yahoo/elide/security/executors/ActivePermissionExecutor.java
@@ -116,6 +116,11 @@ public class ActivePermissionExecutor implements PermissionExecutor {
         };
 
         Function<Expression, ExpressionResult> expressionExecutor = (expression) -> {
+            // for newly created object in PatchRequest limit to User checks
+            if (((RequestScope) resource.getRequestScope()).isMutatingMultipleEntities()
+                    && requestScope.getNewPersistentResources().contains(resource)) {
+                return executeUserChecksDeferInline(annotationClass, expression);
+            }
             return executeExpressions(expression, annotationClass, Expression.EvaluationMode.INLINE_CHECKS_ONLY);
         };
 
@@ -125,6 +130,21 @@ public class ActivePermissionExecutor implements PermissionExecutor {
                 Optional.empty(),
                 expressionSupplier,
                 expressionExecutor);
+    }
+
+    /**
+     * We will only run User checks during patch extension for newly created objects because relationships are not yet
+     * complete. The inline checks may fail (relationships are not yet fixed up). If the user checks return deferred, we
+     * will defer all of the inline checks to commit phase.
+     */
+    private <A extends Annotation> ExpressionResult executeUserChecksDeferInline(Class<A> annotationClass,
+            Expression expression) {
+        ExpressionResult result =
+                executeExpressions(expression, annotationClass, Expression.EvaluationMode.USER_CHECKS_ONLY);
+        if (result == DEFERRED) {
+            commitCheckQueue.add(new QueuedCheck(expression, annotationClass));
+        }
+        return result;
     }
 
     /**
@@ -176,8 +196,11 @@ public class ActivePermissionExecutor implements PermissionExecutor {
         };
 
         Function<Expression, ExpressionResult> expressionExecutor = (expression) -> {
-            commitCheckQueue.add(new QueuedCheck(expression, annotationClass));
-            return ExpressionResult.DEFERRED;
+            if (((RequestScope) resource.getRequestScope()).isMutatingMultipleEntities()
+                    && requestScope.getNewPersistentResources().contains(resource)) {
+                return executeUserChecksDeferInline(annotationClass, expression);
+            }
+            return executeExpressions(expression, annotationClass, Expression.EvaluationMode.INLINE_CHECKS_ONLY);
         };
 
         return checkPermissions(
@@ -344,6 +367,7 @@ public class ActivePermissionExecutor implements PermissionExecutor {
      * @param resourceClass Resource class
      * @return the filter expression for the class, if any
      */
+    @Override
     public Optional<FilterExpression> getReadPermissionFilter(Class<?> resourceClass) {
         FilterExpression filterExpression =
                 expressionBuilder.buildAnyFieldFilterExpression(resourceClass, requestScope);